test: :truck: move db populator logic into its own package so it can be reused in test suite
